meilisearch/playground/requirements.txt
streamlit==1.39.0
meilisearch
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
numpy==1.26.4
pandas==2.2.2
python-dotenv
requests
